added voronoi region function, extracted parts of panoptica pipeline into its own file
